k8s, lbmap, loadbalancer, service: Plumb Maglev table size annotation
Retrieve the Maglev table size value from the annotation on the Service
and pass it down to the LB BPF maps.

+++ pkg/maps/lbmap/maglev.go
@@ -128,7 +128,7 @@ func newInnerMaglevMap(name string) *bpf.Map {
 		name,
 		bpf.MapTypeArray,
 		&MaglevInnerKey{}, int(unsafe.Sizeof(MaglevInnerKey{})),
-		&MaglevInnerVal{}, int(unsafe.Sizeof(uint16(0)))*option.Config.MaglevTableSize,
+		&MaglevInnerVal{}, int(unsafe.Sizeof(uint16(0))*uintptr(option.Config.MaglevTableSize)),
 		1, 0, 0,
 		bpf.ConvertKeyValue,
 		&bpf.NewMapOpts{},
